Crate dr_ones

Drone implementation for the dr_ones network simulator. Provides the Dr_One type which implements the wg_2024::drone::Drone trait.

Logging Control

The drone and other network nodes produce log output that can be controlled globally:

// Disable all node logging
dr_ones::disable_logging();

// Re-enable logging
dr_ones::enable_logging();

Structs

Drone

Implementation of a drone node in the network. Responsible for routing packets and managing network connections.

Functions

disable_logging

Disables logging to stdout.

enable_logging

Enables logging to stdout and resets any log file redirection.

redirect_logs_to_file

Redirects log output to a file.